checkpoint 5: setup the preconditions for group by
